+++ heating_planner/back/cst.py
@@ -1,0 +1,1 @@
+FRANCE_AERA_SQ_KM = 543_940
